+++ src/TodoForm.jsx
@@ -1,0 +1,11 @@
+function TodoForm(){
+    return (
+        <form>
+            <label htmlFor="todoTitle">Todo</label>
+            <input type="text" id="todoTitle"></input>
+            <button>Add Todo</button>
+        </form>
+    )
+}
+
+export default TodoForm
